# Cache slow/computed keys using an LRU