Caching the "slow cache keys" (computed keys based on mounts & copied artifacts) shaves off about 20-30s from my tests builds. I need to confirm that the values I've chosen for the LRU cache keys are reliable, but the builds, especially loading cache layers, are visually faster. It's also much lighter on the CPU and saves around parallelized 100s. 

```
With optimization
earthly-dev -P ./tests+ga-no-qemu-group1 --BUILDKIT_PROJECT=../buildkit  12.59s user 3.64s system 12% cpu 2:13.98 tota
earthly-dev -P ./tests+ga-no-qemu-group1 --BUILDKIT_PROJECT=../buildkit  11.96s user 3.28s system 12% cpu 2:00.92 total

Without
earthly-dev -P ./tests+ga-no-qemu-group1 --BUILDKIT_PROJECT=../buildkit  11.36s user 3.36s system 9% cpu 2:33.95 total
earthly-dev -P ./tests+ga-no-qemu-group1 --BUILDKIT_PROJECT=../buildkit  11.13s user 3.22s system 9% cpu 2:25.74 total
```